+++ src/Move.hpp
@@ -5,8 +5,8 @@
 using Position = uint8_t;
 
 struct Coordinates {
-    uint8_t x;
-    uint8_t y;
+    unsigned int x;
+    unsigned int y;
 };
 
 struct Move {
